Logique de base ennemi + spawner fonnctionnel
- La state machine fonctionne mais que pour le light

- Le spawner ne gère pas encore l'état spawn, à retravailler

- A part ça, le spawner est fonctionnel

- Début du travail sur les waves, besoin de spécifications pour avancer

+++ Assets/Scripts/Enemy/EnemyAttack.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class EnemyAttack : MonoBehaviour //Ce script est juste temporaire pour tester les mécaniques
                                          //Ce genre de script n'est utilse que si l'ennemi n'a qu'une seule attaque
 {
-    [SerializeField] private float attackDamage; //Valeur à changer dans l'inspecteur
+    [SerializeField] private int attackDamage; //Valeur à changer dans l'inspecteur
 
     private void OnCollisionStay2D(Collision2D collision)
     {
